requirements.txt
numpy>=1.24.0
pandas>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
torch>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
tqdm>=4.65.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

